requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
djangorestframework==3.15.1
gunicorn==23.0.0
importlib-metadata==6.7.0
packaging==25.0
Pillow==9.5.0
psycopg2-binary==2.9.10
pyngrok==7.0.5
pytz==2025.2
PyYAML==6.0.1
sqlparse==0.5.3
typing_extensions==4.7.1
tzdata==2025.2
whitenoise==6.9.0
zipp==3.15.0
